feat: 🎸 run jobs killing routines with some random delay (#1364)
* feat: 🎸 run jobs killing routines with some random delay

instead of running the zombie killing routine, and the long jobs killing
routine on a regular basis, we run them between 50% and 150% of that
delay. Fixes #1330

* style: 💄 fix comment position

* fix: 🐛 fix type

* Update services/worker/src/worker/executor.py


* fix: 🐛 use float, not int, for variables that repr. seconds

---------

+++ libs/libcommon/src/libcommon/queue.py
@@ -783,7 +783,7 @@ class Queue:
             return
         job.update(last_heartbeat=get_datetime())
 
-    def get_zombies(self, max_seconds_without_heartbeat: int) -> List[JobInfo]:
+    def get_zombies(self, max_seconds_without_heartbeat: float) -> List[JobInfo]:
         """Get the zombie jobs.
         It returns jobs without recent heartbeats, which means they crashed at one point and became zombies.
         Usually `max_seconds_without_heartbeat` is a factor of the time between two heartbeats.
